bpf: Report original source IP in TRACE_TO_LXC
In order to correlate flow events on multiple nodes, a possible service IP
translation must be known. The TRACE_TO_LXC event is situated in a position to
report this as reverse translation has been performed just before. Extend
the trace_notify structure to include the original source IP if available.

Example:
```
-> endpoint 2614 flow 0x5b9d9e5a identity 56995->31468 state new ifindex lxcc3be4eeac8fe orig-ip 10.16.182.13: 10.16.182.13:33020 -> 10.16.228.151:80 tcp SYN
-> endpoint 3745 flow 0xa4419b6f identity 31468->56995 state reply ifindex lxc155152faaf3f orig-ip 10.16.228.151: 172.20.0.187:80 -> 10.16.182.13:33020 tcp ACK, RST
```

Signed-off-by: Thomas Graf <thomas@cilium.io>