daemon: Run sync-lb controller after bootstrap is finished
Unfortunately, just having "daemon: Fix race condition when syncing svcs
with k8s" is not enough to fix the race.

The problem is that in our case cache.WaitForCacheSync() guarantees that
a service-related k8s resource object was consumed by k8s_service_cache,
but not by k8sServiceHandler() which is responsible for calling
the pkg/service.Service handlers. This is due the following indirection
we have:

    kube-apiserver ->
                      <- cache.Informer()
                           k8s_service_cache ->
                                                 <- k8sServiceHandler()
                                                      Service.UpsertService()

To fix the race, we need to wait until the events have been consumed
by k8sServiceHandler(). To do that, we introduce the waitgroup which
is incremented when calling k8s_service_cache and decremented by
the k8sServiceHandler after calling the Service methods.

Signed-off-by: Martynas Pumputis <m@lambda.lt>