bpf: Repurpose upper bits of cap_len
The cap_len in practice was previously always TRACE_PAYLOAD_LEN if a
notification message is emitted, which by default is 128 and was not
known to exceed 65535 bytes. By repurposing the upper 16 bits of this
field, we can now inject a length for monitor messages to support
extensions of the trace notify header to accommodate new metadata.

Note that this may cause unnecessary dissection when using an older
monitor client with a newer version of Cilium, but in practice the
library that Cilium provides handled this in a reasonable manner in past
versions, and also it's more likely for users to run newer clients with
older Cilium, so this problem is largely ignored for this commit.

Signed-off-by: Joe Stringer <joe@cilium.io>